fix(guest-agent): add explicit type casts to fix seccomp filter test assertions

+++ cmd/guest-agent/sandbox_proc_test.go
@@ -19,25 +19,25 @@ func TestBuildSeccompFilter(t *testing.T) {
 	assert.Len(t, filter, expectedLen)
 
 	// First instruction: load architecture (offset 4 in seccomp_data)
-	assert.Equal(t, bpfLD|bpfW|bpfABS, filter[0].Code, "first instruction should load arch")
+	assert.Equal(t, uint16(bpfLD|bpfW|bpfABS), filter[0].Code, "first instruction should load arch")
 	assert.Equal(t, uint32(4), filter[0].K, "first instruction offset should be 4")
 
 	// Second instruction: check architecture
-	assert.Equal(t, bpfJMP|bpfJEQ|bpfK, filter[1].Code, "second instruction should be arch check jump")
+	assert.Equal(t, uint16(bpfJMP|bpfJEQ|bpfK), filter[1].Code, "second instruction should be arch check jump")
 
 	// Third instruction: load syscall number (offset 0)
-	assert.Equal(t, bpfLD|bpfW|bpfABS, filter[2].Code, "third instruction should load syscall nr")
+	assert.Equal(t, uint16(bpfLD|bpfW|bpfABS), filter[2].Code, "third instruction should load syscall nr")
 	assert.Equal(t, uint32(0), filter[2].K, "third instruction offset should be 0")
 
 	// Second-to-last: ALLOW
 	allowIdx := len(filter) - 2
-	assert.Equal(t, bpfRET|bpfK, filter[allowIdx].Code, "second-to-last should be RET")
-	assert.Equal(t, seccompRetAllow, filter[allowIdx].K, "second-to-last should be ALLOW")
+	assert.Equal(t, uint16(bpfRET|bpfK), filter[allowIdx].Code, "second-to-last should be RET")
+	assert.Equal(t, uint32(seccompRetAllow), filter[allowIdx].K, "second-to-last should be ALLOW")
 
 	// Last: ERRNO(EPERM)
 	lastIdx := len(filter) - 1
-	assert.Equal(t, bpfRET|bpfK, filter[lastIdx].Code, "last should be RET")
-	assert.Equal(t, seccompRetErrno|errnoEPERM, filter[lastIdx].K, "last should be ERRNO(EPERM)")
+	assert.Equal(t, uint16(bpfRET|bpfK), filter[lastIdx].Code, "last should be RET")
+	assert.Equal(t, uint32(seccompRetErrno|errnoEPERM), filter[lastIdx].K, "last should be ERRNO(EPERM)")
 }
 
 func TestBlockedSyscalls(t *testing.T) {
@@ -47,7 +47,7 @@ func TestBlockedSyscalls(t *testing.T) {
 
 	switch runtime.GOARCH {
 	case "amd64":
-		assert.Equal(t, auditArchX86_64, auditArch)
+		assert.Equal(t, uint32(auditArchX86_64), auditArch)
 		expected := []uint32{
 			sysProcessVMReadvAmd64,
 			sysProcessVMWritevAmd64,
@@ -59,7 +59,7 @@ func TestBlockedSyscalls(t *testing.T) {
 			assert.Equal(t, nr, blocked[i], "blocked[%d]", i)
 		}
 	case "arm64":
-		assert.Equal(t, auditArchAARCH64, auditArch)
+		assert.Equal(t, uint32(auditArchAARCH64), auditArch)
 	}
 }
